services/ai-service/pyproject.toml
[project]
name = "jobpilot-ai-service"
version = "1.0.0"
description = "AI Service for JobPilot - Job matching and resume optimization"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "openai==1.10.0",
    "anthropic==0.18.0",
    "pinecone-client==3.0.0",
    "numpy==1.26.3",
    "scikit-learn==1.4.0",
    "redis==5.0.1",
    "httpx==0.26.0",
    "python-multipart==0.0.6",
    "python-jose[cryptography]==3.3.0",
    "tenacity==8.2.3",
    "structlog==24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.12.1",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
